Тесты для мапперов

+++ src/main/kotlin/pro/liga/data/player/rating/RatingDTO.kt
@@ -5,6 +5,6 @@ import pro.liga.data.DTO
 
 data class RatingDTO(
     val playerId: Int,
-    val rating: Int,
+    val rating: Short,
     val date: LocalDate
 ): DTO
